Begin WIP VPS configuration
Thus begins the process. This commit should be removed eventually.

I have a very low power and low memory VPS. It's not able to build itself
in any reasonable amount of time. I do have an x86 machine sitting around
but I'd rather not turn it on if I don't have to because it's hot and
slow. The best solution is to use a virtual machine on my ARM Darwin
machines. linux-builder already exists, but it uses QEMU, which is slow
at x86 emulation. On macOS, you can use Rosetta 2 to translate
instructions to ARM, but this isn't available in QEMU. You could either
run a UTM VM (which has its own set of limitations) or use some other
kind of VM runner that has support for Virtualisation.Framework.

I've opted for using that other solution, specifically Lima. I didn't
want to use Docker (for a variety of reasons), so this seems like a good
alternative.

Now, bootstrapping. Each commit, GitHub will build and publish a VM image
that can be downloaded to create the VM. From that point on, the VM can
build itself.